pyproject.toml
[project]
name = "themind"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>=1.91.0",
    "pytest>=8.4.1",
    "python-dotenv>=1.1.1",
    "PyYAML>=6.0.1",
]

[tool.setuptools.packages.find]
where = ["."]
